Update template (#2)
This template hasn't been updated pretty much since I created it, and a
great deal has changed with the repo since then.

    Added playwright and typecheck github actions
    Some updated gitignores
    Reorganized the clients folder
    Reorganized the deploy folder
    Cleaned up the notes folder
    Updated the top-level test scripts
    Added an identity service

There's still more in this template than I'd like there to be; ideally
it would be pretty empty and one could use workflows to bootstrap
everything here. The main blocker to that is to improve the docker
situation, and by that I mean probably replacing it with Nix. Need to
set aside some time there, but once that's done I think the proper way
to structure my clients will also follow from that. Until then, anything
that involves changing things in deploy/ is going to be fairly manual.